+++ tests/jest/aggregate-themes.test.js
@@ -1,10 +1,11 @@
+const path = require('path');
 const ThemeAggregator = require('../../scripts/aggregate-themes/theme-aggregator');
 
 global.console = { log: jest.fn(), warn: jest.fn() };
 
 describe('Theme Aggregator', () => {
   describe('aggregate', () => {
-    it('returns null if there is no terra-theme-config.js', () => {
+    it('returns null if there is no terra-theme-config.js and no config passed by env var.', () => {
       const file = ThemeAggregator.aggregate();
 
       expect(file).toBeNull();
@@ -12,20 +13,31 @@ describe('Theme Aggregator', () => {
   });
 
   describe('aggregateTheme', () => {
-    it('returns an array of aggregated default theme files', () => {
+    it('returns an array of aggregated default theme file paths', () => {
       const options = { theme: 'terra-mock-dark-theme' };
 
       const files = ThemeAggregator.aggregateTheme('terra-mock-dark-theme', options);
-      const expected = ['./tests/jest/fixtures/themes/terra-mock-dark-theme/root-theme.scss'];
+      const expected = ['../tests/jest/fixtures/themes/terra-mock-dark-theme/root-theme.scss'];
 
       expect(files).toEqual(expected);
     });
 
-    it('returns an array of aggregated scoped theme files', () => {
+    it('returns an array of aggregated scoped theme files paths', () => {
       const options = { scoped: ['terra-mock-dark-theme'] };
 
       const files = ThemeAggregator.aggregateTheme('terra-mock-dark-theme', options);
-      const expected = ['./tests/jest/fixtures/themes/terra-mock-dark-theme/scoped-theme.scss'];
+      const expected = ['../tests/jest/fixtures/themes/terra-mock-dark-theme/scoped-theme.scss'];
+
+      expect(files).toEqual(expected);
+    });
+  });
+
+  describe('findThemeVariableFiles', () => {
+    it("returns the requested theme's files", () => {
+      const themeName = 'terra-mock-dark-theme';
+
+      const files = ThemeAggregator.findThemeVariableFiles(themeName);
+      const expected = ['tests/jest/fixtures/themes/terra-mock-dark-theme/terra-mock-dark-theme.scss'];
 
       expect(files).toEqual(expected);
     });
@@ -77,14 +89,16 @@ describe('Theme Aggregator', () => {
   describe('resolve', () => {
     it('resolves a file path from the node_modules directory', () => {
       const resolvedFile = ThemeAggregator.resolve('node_modules/terra-mock-dark-theme/root-file.scss');
+      const expected = 'terra-mock-dark-theme/root-file.scss';
 
-      expect(resolvedFile).toEqual('terra-mock-dark-theme/root-file.scss');
+      expect(resolvedFile).toEqual(expected);
     });
 
     it('resolves a relative file path', () => {
       const resolvedFile = ThemeAggregator.resolve('fixtures/themes/terra-mock-dark-theme/root-file.scss');
+      const expected = '../fixtures/themes/terra-mock-dark-theme/root-file.scss';
 
-      expect(resolvedFile).toEqual('./fixtures/themes/terra-mock-dark-theme/root-file.scss');
+      expect(resolvedFile).toEqual(expected);
     });
   });
 
@@ -114,6 +128,21 @@ describe('Theme Aggregator', () => {
 
       // eslint-disable-next-line no-console
       expect(console.warn).toHaveBeenCalled();
+    });
+  });
+
+  describe('writeSCSSFile', () => {
+    it('returns the generated SCSS file path relative to the given output path.', () => {
+      const theme = 'test-theme';
+
+      const outputPath = path.resolve(process.cwd(), 'tests', 'jest', 'fixtures', 'generatedThemes');
+
+      const scssFile = ThemeAggregator.writeSCSSFile({
+        assets: [theme], themeName: theme, prefix: 'scoped', scopeSelector: theme, outputPath,
+      });
+
+      const expected = ['./scoped-test-theme.scss'];
+      expect(scssFile).toEqual(expected);
     });
   });
 });