+++ CHANGELOG.md
@@ -1,10 +1,31 @@
 Changelog
 =========
 
-Unreleased
+5.7.1 - (September 12, 2019)
+----------
+### Fixed
+* Fixed issue with theme aggregation where the default theme wasn't honored
+
+5.7.0 - (September 12, 2019)
 ----------
 ### Added
 * Added a post-install console warning if terra-toolkit is included as a hard dependency
+* Fallback theme generation strategy for theme aggregation.
+* Theme visual regression support.
+
+### Changed
+* Updated webpack range from "webpack": "^4.30.0" to "webpack": ">=4.30.0 <4.40.0" to avoid consuming buggy behavior in webpack 4.40.0
+
+5.6.2 - (September 6, 2019)
+----------
+### Fixed
+* Reverted post-install console warning if terra-toolkit is included as a hard dependency
+
+5.6.1 - (September 5, 2019)
+----------
+### Added
+* Added a post-install console warning if terra-toolkit is included as a hard dependency
+* Added a line to webpack config to only resolve to 'main'
 
 5.6.0 - (August 30, 2019)
 ----------